+++ custom-sign-in/CustomSignIn/app/ProfileScreen.js
@@ -1,0 +1,82 @@
+﻿/*
+ * Copyright (c) 2019, Okta, Inc. and/or its affiliates. All rights reserved.
+ * The Okta software accompanied by this notice is provided pursuant to the Apache License, Version 2.0 (the "License.")
+ *
+ * You may obtain a copy of the License at http://www.apache.org/licenses/LICENSE-2.0.
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
+ * WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ *
+ * See the License for the specific language governing permissions and limitations under the License.
+ */
+
+import React, {Fragment} from 'react';
+import {
+  SafeAreaView,
+  ScrollView,
+  Button,
+  StyleSheet,
+  Text,
+  View,
+  StatusBar,
+  TextInput,
+  ActivityIndicator,
+} from 'react-native';
+
+export default class ProfileScreen extends React.Component {
+  static navigationOptions = {
+    title: 'Details',
+  };
+  render() {
+    const {navigation} = this.props;
+    const transaction = navigation.getParam('transaction', 'NO-ID');
+    //const name = ((user || {}).personalInfo || {}).name;
+    const login = transaction.data._embedded.user.id;
+    return (
+      <Fragment>
+        <StatusBar barStyle="dark-content" />
+        <SafeAreaView style={styles.container}>
+        <Text>{login}</Text>
+        </SafeAreaView>
+      </Fragment>
+    );
+  }
+}
+
+const styles = StyleSheet.create({
+  textInput: {
+    marginTop: 10,
+    height: 40,
+    borderColor: 'gray',
+    borderWidth: 1,
+  },
+  buttonContainer: {
+    flexDirection: 'column',
+    justifyContent: 'space-between',
+    marginTop: 10,
+  },
+  button: {
+    borderRadius: 40,
+    width: 200,
+    height: 40,
+    marginTop: 40,
+    marginBottom: 10,
+    marginHorizontal: 10,
+  },
+  container: {
+    flex: 1,
+    flexDirection: 'column',
+    backgroundColor: '#FFFFFF',
+    alignItems: 'center',
+  },
+  context: {
+    marginTop: 20,
+  },
+  title: {
+    fontSize: 30,
+    fontWeight: 'bold',
+    color: '#0066cc',
+    paddingTop: 40,
+    textAlign: 'center',
+  },
+});
